howtosayiloveyou
Using if else statements created a program to ask what language you would like to say "I love you" in and respond in that language. To continue the program I used a while loop to continue to ask about other languages. The program is not perfected as I am just beginning C++ programming.